+++ cmd/boxie/internal/commands/create/create.go
@@ -1,21 +1,21 @@
-// Package get is an entry point and help tools for get commands
-package get
+// Package create is an entry point and help tools for create commands
+package create
 
 import (
 	"github.com/spf13/cobra"
 	"github.com/twelvee/boxie/internal/boxie/handlers/get"
 )
 
-// NewGetCommand is get command entry point
-func NewGetCommand() *cobra.Command {
+// NewCreateCommand is create command entry point
+func NewCreateCommand() *cobra.Command {
 	var (
 		command   *cobra.Command
 		namespace string
 	)
 	command = &cobra.Command{
-		Use:     "get",
-		Short:   "Get a list of saved resources",
-		Long:    "Get a list of earlier created resources. Use requires the resource type as the first argument.",
+		Use:     "create",
+		Short:   "Create and save the resource",
+		Long:    "Create and save the resource that you are going to use when creating environments.",
 		Example: getExample(),
 		Args:    cobra.MinimumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
