+++ cmd/boxie/internal/commands/delete/delete.go
@@ -3,7 +3,7 @@ package delete
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/twelvee/boxie/internal/boxie/handlers"
+	"github.com/twelvee/boxie/internal/boxie/handlers/delete"
 )
 
 // NewDeleteCommand is delete command entry point
@@ -19,7 +19,7 @@ func NewDeleteCommand() *cobra.Command {
 		Example: getExample(),
 		Args:    cobra.MinimumNArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			handlers.HandleDeleteCommand(command.Context(), args[0], args[1], namespace)
+			delete.HandleDeleteCommand(command.Context(), args[0], args[1], namespace)
 			return nil
 		},
 	}
