+++ cmd/boxie/internal/commands/get/get_helper.go
@@ -1,3 +1,4 @@
+// Package get is an entry point and help tools for get commands
 package get
 
 func getExample() string {
