+++ internal/boxie/handlers/k8sExecutable.go
@@ -7,7 +7,7 @@ import (
 	"github.com/twelvee/boxie/internal/boxie"
 )
 
-// HandleGetCommand prepare all data to use k8s cluster with provided flags
+// KuberExecutable prepare all data to use k8s cluster with provided flags
 func KuberExecutable(context context.Context, namespace string) {
 	err := boxie.GetEnvironmentService().PrepareToWorkWithNamespace(namespace)
 	if err != nil {
